+++ code/pet.store.spring.dao/src/main/java/pet/store/spring/dao/model/interfaces/PetDaoEntityI.java
@@ -1,8 +1,8 @@
 package pet.store.spring.dao.model.interfaces;
 
 public interface PetDaoEntityI {
-	public int getId();
-	public void setId(int id);
+	public long getId();
+	public void setId(long id);
 	public String getName();
 	public void setName(String name);
 	public String getStatus();
