+++ code/pet.store.spring.dao/src/main/java/pet/store/spring/dao/repositories/PetsDaoRepositoryI.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pet.store.spring.dao.model.classes.PetDaoEntityC;
 
 @Repository
-public interface PetsDaoRepositoryI extends PagingAndSortingRepository<PetDaoEntityC, Integer> {
+public interface PetsDaoRepositoryI extends PagingAndSortingRepository<PetDaoEntityC, Long> {
 }
